Module elasticlunr::config

These types are not used for generating Indexes. They are provided to help with creating compatible JSON structures for configuring the JavaScript search function.

Reference: http://elasticlunr.com/docs/configuration.js.html

Structs

SearchOptions

The search configuration map which is passed to the elasticlunr.js Index.search() function.

SearchOptionsField

Used to set the search configuration for a specific field. When expand or bool is None, elasticlunr.js will use the value from the global configuration. The boost field, if present, increases the importance of this field when ordering search results.

Enums

SearchBool

Sets which boolean model is used for searching with multiple terms. Defaults to Or.