Enum elasticlunr::lang::Language

pub enum Language {
    Danish,
    Dutch,
    English,
    Finnish,
    French,
    German,
    Italian,
    Portuguese,
    Romanian,
    Russian,
    Spanish,
    Swedish,
    Turkish,
    // some variants omitted
}

Used to configure the Index for a specific lanugage.

Variants
DanishDutchEnglishFinnishFrenchGermanItalianPortugueseRomanianRussianSpanishSwedishTurkish
Methods

impl Language

pub fn from_code(code: &str) -> Option<Language>

Returns the Language for the given two-character ISO 639-1 language code if the language is supported. Returns None if not supported.

Note:

The ISO 639-1 code for Dutch is "nl". However "du" is used for the module name and pipeline suffix in order to match lunr-languages.

pub fn to_code(&self) -> &'static str

Returns the two-character ISO 639-1 language code for the Language.

Note:

The ISO 639-1 code for Dutch is "nl". However "du" is used for the module name and pipeline suffix in order to match lunr-languages.