elasticlunr

Module lang

Intended to be compatible with https://github.com/MihaiValentin/lunr-languages. Each supported language has a trimmer, a stop word filter, and a stemmer. Most users will not need to use these modules directly.

Modules

common

Structs

English

Traits

Language

Functions

from_code

Returns the Language for the given two-character ISO 639-1 language code if the language is supported. Returns None if not supported.

from_name

Returns the Language for the given English language name if the language is supported. Returns None if not supported. The first letter must be capitalized.

languages

Returns a list of all the Language implementations in the crate

tokenize_whitespace

Splits a text string into a vector of individual tokens.