Module elasticlunr::lang

Intended to be compatible with https://github.com/MihaiValentin/lunr-languages. Each supported language has a trimmer, a stop word filter, and a stemmer. Most users will not need to use these modules directly.

Modules

da

de

du

en

es

fi

fr

it

pt

ro

ru

sv

tr

Structs

LanguageIter

Enums

Language
Used to configure the Index for a specific lanugage.