Struct elasticlunr::IndexBuilder

pub struct IndexBuilder { /* fields omitted */ }

A builder for an Index with custom parameters.

Methods

impl IndexBuilder

pub fn new() -> Self

pub fn save_docs(self, save: bool) -> Self

Set whether or not documents should be saved in the Index.

pub fn add_field(self, field: &str) -> Self

Add a document field to the Index.

pub fn set_ref(self, ref_field: &str) -> Self

Set the key used to store the document reference field.

pub fn build(self) -> Index

Build an Index from this builder.